# doc/openapi.yaml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: Bildungsportal
...
security:
    - bearerAuth: []
paths:
...
  /education/{id}:
    get:
      summary: "Returns specific education"
      operationId: "singleEducation"
      tags:
        - education
      security: []
      parameters:
        - in: path
          name: id
          description: "Id of education"
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Education'
        "404":
          description: "Education not found"
    patch:
      summary: "Updates an education"
      operationId: "updateEducationOfLocation"
      tags:
        - education
      parameters:
        - in: path
          name: id
          description: "Id of education"
          required: true
          schema:
            type: integer
      requestBody:
        description: "Updated education"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EducationProperties'
      responses:
        "200":
          description: "Education updated"
        "400":
          description: "Request is incomplete or invalid"
        "401":
          description: "User is not authorized"
        "403":
          description: "Insufficient permissions"
        "404":
          description: "Education not found"
...
  /matriculation/{id}:
    patch:
      summary: "Updates an matriculation"
      operationId: "updateMatriculation"
      tags:
        - matriculation
      parameters:
        - in: path
          name: id
          description: "Id of matriculation"
          required: true
          schema:
            type: integer
      requestBody:
        description: "Updated matriculation"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatriculationProperties'
      responses:
        "200":
          description: "Matriculation updated"
        "400":
          description: "Request is incomplete or invalid"
        "401":
          description: "User is not authorized"
        "403":
          description: "Insufficient permissions"
        "404":
          description: "Matriculation not found"
  /education/{id}/matriculation:
    get:
      summary: "Returns all matriculation of specific education"
      operationId: "allMatriculationsOfEducation"
      tags:
        - matriculation
      parameters:
        - in: path
          name: id
          description: "Id of education"
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/Matriculation'
    post:
      summary: "Creates a new matriculation for specific education"
      operationId: "createMatriculationOfEducation"
      tags:
        - matriculation
      parameters:
        - in: path
          name: id
          description: "Id of education"
          required: true
          schema:
            type: integer
      requestBody:
        description: "Student id for matriculation"
        required: true
        content:
          application/json:
            schema:
              type: "integer"
      responses:
        "201":
          description: "Matriculation created"
        "400":
          description: "Request is incomplete or invalid"
        "401":
          description: "User is not authorized"
        "403":
          description: "Insufficient permissions"
  /education/{id}/assessment:
    get:
      summary: "Returns all assessment of specific education"
      operationId: "allAssessmentsOfEducation"
      tags:
        - assessment
      parameters:
        - in: path
          name: id
          description: "Id of education"
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Successful operation"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: '#/components/schemas/Assessment'
    post:
      summary: "Creates a new assessment for specific education"
      operationId: "createAssessmentOfEducation"
      tags:
        - assessment
      parameters:
        - in: path
          name: id
          description: "Id of education"
          required: true
          schema:
            type: integer
      requestBody:
        description: "New assessment to be created"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Assessment'
      responses:
        "201":
          description: "Assessment created"
        "400":
          description: "Request is incomplete or invalid"
        "401":
          description: "User is not authorized"
        "403":
          description: "Insufficient permissions"
  /assessment/{id}:
    patch:
      summary: "Updates an assessment"
      operationId: "updateAssessment"
      tags:
        - assessment
      parameters:
        - in: path
          name: id
          description: "Id of assessment"
          required: true
          schema:
            type: integer
      requestBody:
        description: "Updated assessment"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentProperties'
      responses:
        "200":
          description: "Assessment updated"
        "400":
          description: "Request is incomplete or invalid"
        "401":
          description: "User is not authorized"
        "403":
          description: "Insufficient permissions"
        "404":
          description: "Assessment not found"
  /education/{id}/qualification:
    get:
      summary: "Returns all possible qualifications of specific education"
      operationId: "allQualificationsOfEducation"
      tags:
        - qualification
      security: []
      parameters:
        - in: path
          name: id
          description: "Id of education"
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "List of qualification ids"
          content:
            application/json:
              schema:
                type: "array"
                items:
                  type: "integer"
    post:
      summary: "Adds a possible qualification for specific education"
      operationId: "addQualificationsOfEducation"
      tags:
        - requirement
      parameters:
        - in: path
          name: id
          description: "Id of education"
          required: true
          schema:
            type: integer
      requestBody:
        description: "List of qualification ids"
        required: true
        content:
          application/json:
            schema:
              type: "array"
              items:
                type: "integer"
      responses:
        "201":
          description: "Qualification added"
        "400":
          description: "Request is incomplete or invalid"
        "401":
          description: "User is not authorized"
        "403":
          description: "Insufficient permissions"
  /education/{educationId}/qualification/{qualificationId}:
    delete:
      summary: "Removes a possible qualifications for specific education"
      operationId: "removeQualificationsOfEducation"
      tags:
        - requirement
      parameters:
        - in: path
          name: educationId
          description: "Id of education"
          required: true
          schema:
            type: integer
        - in: path
          name: qualificationId
          description: "Id of qualification"
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "Qualification removed"
        "404":
          description: "Requirement does not exist"
...
components:
  schemas:
...
    EducationProperties:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          readOnly: true
        locationId:
          type: "integer"
          format: "int64"
        title:
          type: "string"
          format: "string"
        description:
          type: "string"
          format: "string"
    Education:
      type: "object"
      allOf:
        - $ref: '#/components/schemas/EducationProperties'
        - required:
            - title
            - description
...
    MatriculationProperties:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          readOnly: true
        studentId:
          type: "integer"
          format: "int64"
        educationId:
          type: "integer"
          format: "int64"
        date:
          type: "string"
          format: "iso8601"
    Matriculation:
      type: "object"
      allOf:
        - $ref: '#/components/schemas/MatriculationProperties'
        - required:
            - date
    AssessmentProperties:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          readOnly: true
        title:
          type: "string"
          format: "string"
        educationId:
          type: "integer"
          format: "int64"
          readOnly: true
    Assessment:
      type: "object"
      allOf:
        - $ref: '#/components/schemas/AssessmentProperties'
        - required:
            - title
...
  securitySchemes:
    bearerAuth:
        type: http
        scheme: bearer
        bearerFormat: JWT